AG-17828 Move dev-time validation overlay docs and e2e coverage to their own page
Relocate the validation overlay examples, docs prose, and e2e coverage from
the Overlays page to a new Dev-Time Validation page, restoring the prose that
was intentionally left out when the examples/e2e were previously restored.
