+++ packages/ag-charts-website/e2e/dev-validation.spec.ts-snapshots/validation-overlay-dark-chromium-linux.png
