+++ packages/ag-charts-website/e2e/dev-validation.spec.ts-snapshots/validation-overlay-multi-chromium-linux.png
