+++ packages/ag-charts-website/e2e/dev-validation.spec.ts-snapshots/validation-overlay-warning-chromium-linux.png
