+++ packages/ag-charts-website/src/content/docs/dev-validation/_examples/validation-overlay-multi/styles.css
